moving server connection code from transport to core
