+++ bundle-core/pom.xml
@@ -84,11 +84,6 @@
             <version>1.3.5</version>
             <optional>true</optional>
         </dependency>
-        <!--<dependency>
-            <groupId>net.devh</groupId>
-            <artifactId>grpc-server-spring-boot-starter</artifactId>
-            <version>2.15.0.RELEASE</version>
-        </dependency>-->
     </dependencies>
 
     <build>
