+++ bundle-core/src/main/java/net/discdd/utils/FileUtils.java
@@ -1,10 +1,24 @@
 package net.discdd.utils;
 
+import com.google.protobuf.ByteString;
+import net.discdd.bundletransport.service.BundleDownloadResponse;
+
+import java.io.File;
 import java.io.IOException;
+import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.nio.file.StandardOpenOption;
+import java.util.Objects;
+import java.util.logging.Logger;
+
+import static java.util.logging.Level.INFO;
+import static java.util.logging.Level.WARNING;
 
 public class FileUtils {
+    private static final Logger logger = Logger.getLogger(FileUtils.class.getName());
+
     public static void createFileWithDefaultIfNeeded(Path path, byte[] bytes) throws IOException {
         if (path.toFile().exists()) return;
         var parent = path.getParent().toFile();
@@ -16,5 +30,57 @@ public class FileUtils {
 
     public static void createEmptyFileIfNeeded(Path path) throws IOException {
         createFileWithDefaultIfNeeded(path, new byte[0]);
+    }
+
+    public static String getFileNamesListFromDirectory(String directory) {
+        String fileList = "";
+        File dir = new File(directory);
+        File[] Files = dir.listFiles();
+        if (Files != null && Files.length > 0) {
+            for (File file : Files) {
+                fileList = fileList + "," + file.getName();
+            }
+        }
+        return fileList;
+    }
+
+    public static OutputStream getFilePath(BundleDownloadResponse response, String Receive_Directory) throws IOException {
+        String fileName = response.getMetadata().getBid();
+        File directoryReceive = new File(Receive_Directory + File.separator + response.getMetadata().getTransportId());
+        if (!directoryReceive.exists()) {
+            directoryReceive.mkdirs();
+        }
+
+        return Files.newOutputStream(Paths.get(
+                Receive_Directory + File.separator + response.getMetadata().getTransportId() + File.separator +
+                        fileName), StandardOpenOption.CREATE, StandardOpenOption.APPEND);
+    }
+
+    public static void writeFile(OutputStream writer, ByteString content) throws IOException {
+        try {
+            writer.write(content.toByteArray());
+            writer.flush();
+        } catch (Exception e) {
+            logger.log(WARNING, "writeFile: " + e.getMessage());
+        }
+
+    }
+
+    public static void closeFile(OutputStream writer) {
+        try {
+            writer.close();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    public static void deleteBundlesFromDirectory(String directory) {
+        File deleteDir = new File(directory);
+        if (deleteDir.listFiles() != null) {
+            for (File bundle : Objects.requireNonNull(deleteDir.listFiles())) {
+                boolean result = bundle.delete();
+                logger.log(INFO, bundle.getName() + "deleted:" + result);
+            }
+        }
     }
 }
